adding sequence observation for transformer

+++ fancy_gym/envs/classic_control/crowd_navigation/l_shape_env.py
@@ -135,13 +135,13 @@ class LShapeCrowdNavigationEnv(BaseCrowdNavigationEnv):
                 rel_crowd_poss if self.n_crowd > 1 else [rel_crowd_poss],
                 [self._agent_vel],
                 dist_walls
-            ]).astype(np.float32).flatten()
+            ]).flatten()
         else:
             return np.concatenate([
                 [rel_goal_pos],
                 [self._agent_vel],
                 dist_walls
-            ]).astype(np.float32).flatten()
+            ]).flatten()
